// src/pages/User/AddUser.d.ts
export interface AddUserParams {
    userName: string;
    userEmail:string;
    userMobileNumber: string;
    userAddress: string;
    userAddress1: string;
    userAddress2: string;
    userAddress3: string;
    officeBuildingName:string;
    officeAreaName:string;
    officeLandLineNumber:string;
    officeAddress1:string;
    officeAddress2:string;
    officePOBoxNumber:string;
}